# collections/SwaggerCollections.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Bytes4All",
        "description": "RestFul API of the Bytes4All application. The API allows users to upload and download e-books",
        "termsOfService": "https://edielsonassis.com.br/termos-a-serem-definidos",
        "contact": {
            "name": "Edielson Assis",
            "email": "grizos.ed@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://github.com/edielson-assis/bytes4all-api/blob/main/LICENSE"
        },
        "version": "v1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/auth/signup": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Adds a new User",
                "description": "Adds a new User by passing in a JSON, XML or YML representation of the user!",
                "operationId": "signup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSignupRequest"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSignupRequest"
                            }
                        },
                        "application/x-yaml": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSignupRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            },
                            "application/x-yaml": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Something is wrong with the request"
                    },
                    "500": {
                        "description": "Internal Server Error - Server error"
                    }
                }
            }
        },
        "/api/v1/auth/signin": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticates a user",
                "description": "Authenticates a user and returns a token",
                "operationId": "signin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSigninRequest"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSigninRequest"
                            }
                        },
                        "application/x-yaml": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSigninRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Authenticated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenAndRefreshTokenResponse"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenAndRefreshTokenResponse"
                                }
                            },
                            "application/x-yaml": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenAndRefreshTokenResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Something is wrong with the request"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid email or password"
                    },
                    "500": {
                        "description": "Internal Server Error - Server error"
                    }
                }
            }
        },
...
        "/api/v1/auth/refresh/{email}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Updates a Token",
                "description": "Refresh token for authenticated user and returns a token",
                "operationId": "refreshToken",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            },
                            "application/x-yaml": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - Something is wrong with the request"
                    },
                    "403": {
                        "description": "Forbidden - Authentication problem"
                    },
                    "404": {
                        "description": "Not found - User not found"
                    },
                    "500": {
                        "description": "Internal Server Error - Server error"
                    }
                }
            }
        },
...
        "/api/v1/auth/{email}": {
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Deletes a User",
                "description": "Deletes a User by their Email",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted user"
                    },
                    "403": {
                        "description": "Forbidden - Authentication problem"
                    },
                    "404": {
                        "description": "Not found - User not found"
                    },
                    "500": {
                        "description": "Internal Server Error - Server error"
                    }
                },
                "security": [
                    {
                        "bearer-key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserResponse": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "description": "Unique identifier of the user",
                        "format": "int64",
                        "example": 1
                    },
                    "fullName": {
                        "type": "string",
                        "description": "Full name of the person",
                        "example": "Robert Martin"
                    },
                    "email": {
                        "type": "string",
                        "description": "Email of the user.",
                        "example": "robert@example.com"
                    }
                },
                "description": "Represents a User"
            },
            "UserSignupRequest": {
                "required": [
                    "email",
                    "fullName",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "fullName": {
                        "maxLength": 2147483647,
                        "minLength": 3,
                        "pattern": "^[a-zA-ZÀ-ÿ\\s']+$",
                        "type": "string",
                        "description": "Full name of the person",
                        "example": "Robert Martin"
                    },
                    "email": {
                        "maxLength": 100,
                        "type": "string",
                        "description": "Email of the user.",
                        "example": "robert@example.com"
                    },
                    "password": {
                        "maxLength": 255,
                        "type": "string",
                        "description": "Password for the user account.",
                        "example": "372@RfI5n&Ml"
                    }
                },
                "description": "Represents the User's Signup"
            },
            "TokenAndRefreshTokenResponse": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string",
                        "description": "Access token provided for authentication",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                    },
                    "refreshToken": {
                        "type": "string",
                        "description": "Refresh token provided for re-authentication",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                    }
                },
                "description": "Representation of a JWT Token and Refresh token"
            },
            "UserSigninRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "maxLength": 100,
                        "type": "string",
                        "description": "Email of the user.",
                        "example": "robert@example.com"
                    },
                    "password": {
                        "maxLength": 255,
                        "type": "string",
                        "description": "Password for the user account.",
                        "example": "372@RfI5n&Ml"
                    }
                },
                "description": "Represents the User's Signin"
            },
            "TokenResponse": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string",
                        "description": "Access token provided for authentication",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                    }
                },
                "description": "Representation of a JWT Token"
            },
...
        },
        "securitySchemes": {
            "bearer-key": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}